add event card and update cards display
